+++ ngmix/admom/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+
+from . import admom
+from .admom import *
+
+from . import admom_nb