+++ ngmix/metacal/metacal.py
@@ -642,10 +642,10 @@ class MetacalFitGaussPSF(MetacalGaussPSF):
         """
         import galsim
 
-        from ..admom import Admom
+        from ..admom import AdmomFitter
         from ..guessers import GMixPSFGuesser, SimplePSFGuesser
         from ..runners import run_psf_fitter
-        from ..fitting import LM
+        from ..fitting import Fitter
 
         psfobs = self.obs.psf
 
@@ -653,7 +653,7 @@ class MetacalFitGaussPSF(MetacalGaussPSF):
         guesser = GMixPSFGuesser(rng=self.rng, ngauss=1)
 
         # try adaptive moments first
-        fitter = Admom(rng=self.rng)
+        fitter = AdmomFitter(rng=self.rng)
 
         res = run_psf_fitter(obs=psfobs, fitter=fitter, guesser=guesser, ntry=ntry)
 
@@ -668,7 +668,7 @@ class MetacalFitGaussPSF(MetacalGaussPSF):
                 'xtol': 1.0e-05,
             }
 
-            fitter = LM(model='gauss', fit_pars=lm_pars)
+            fitter = Fitter(model='gauss', fit_pars=lm_pars)
             guesser = SimplePSFGuesser(rng=self.rng)
 
             res = run_psf_fitter(