+++ ngmix/util.py
@@ -51,3 +51,31 @@ def format_pars(pars, fmt="%8.3g"):
     """
     fmt = " ".join([fmt + " "] * len(pars))
     return fmt % tuple(pars)
+
+
+def get_ratio_var(a, b, var_a, var_b, cov_ab):
+    """
+    get (a/b)**2 and variance in mean of (a/b)
+    """
+
+    if b == 0:
+        raise ValueError("zero in denominator")
+
+    rsq = (a/b)**2
+
+    var = rsq * (var_a/a**2 + var_b/b**2 - 2*cov_ab/(a*b))
+    return var
+
+
+def get_ratio_error(a, b, var_a, var_b, cov_ab):
+    """
+    get a/b and error on a/b
+    """
+    from math import sqrt
+
+    var = get_ratio_var(a, b, var_a, var_b, cov_ab)
+
+    if var < 0:
+        var = 0
+    error = sqrt(var)
+    return error